requirements.txt
streamlit
openai>=1.0.0
python-dotenv
Pillow